Move the clap dependency outside the library
