+++ src/bin/common.rs
@@ -1,7 +1,96 @@
 use y4m;
 use rav1e::*;
+use std::fs::File;
+use std::io;
 use std::io::prelude::*;
 use std::slice;
+use clap::{App, Arg};
+
+pub struct EncoderIO {
+    pub input: Box<Read>,
+    pub output: Box<Write>,
+    pub rec: Option<Box<Write>>,
+}
+
+pub trait FromCli {
+    fn from_cli() -> (EncoderIO, EncoderConfig);
+}
+
+impl FromCli for EncoderConfig {
+    fn from_cli() -> (EncoderIO, EncoderConfig) {
+        let matches = App::new("rav1e")
+            .version("0.1.0")
+            .about("AV1 video encoder")
+           .arg(Arg::with_name("INPUT")
+                .help("Uncompressed YUV4MPEG2 video input")
+                .required(true)
+                .index(1))
+            .arg(Arg::with_name("OUTPUT")
+                .help("Compressed AV1 in IVF video output")
+                .short("o")
+                .long("output")
+                .required(true)
+                .takes_value(true))
+            .arg(Arg::with_name("RECONSTRUCTION")
+                .short("r")
+                .takes_value(true))
+            .arg(Arg::with_name("LIMIT")
+                .help("Maximum number of frames to encode")
+                .short("l")
+                .long("limit")
+                .takes_value(true)
+                .default_value("0"))
+            .arg(Arg::with_name("QP")
+                .help("Quantizer (0-255)")
+                .long("quantizer")
+                .takes_value(true)
+                .default_value("100"))
+            .arg(Arg::with_name("SPEED")
+                .help("Speed level (0(slow)-10(fast))")
+                .short("s")
+                .long("speed")
+                .takes_value(true)
+                .default_value("3"))
+            .arg(Arg::with_name("TUNE")
+                .help("Quality tuning (Will enforce partition sizes >= 8x8)")
+                .long("tune")
+                .possible_values(&Tune::variants())
+                .default_value("psnr")
+                .case_insensitive(true))
+            .get_matches();
+
+
+        let io = EncoderIO {
+            input: match matches.value_of("INPUT").unwrap() {
+                "-" => Box::new(io::stdin()) as Box<Read>,
+                f => Box::new(File::open(&f).unwrap()) as Box<Read>
+            },
+            output: match matches.value_of("OUTPUT").unwrap() {
+                "-" => Box::new(io::stdout()) as Box<Write>,
+                f => Box::new(File::create(&f).unwrap()) as Box<Write>
+            },
+            rec: matches.value_of("RECONSTRUCTION").map(|f| {
+                Box::new(File::create(&f).unwrap()) as Box<Write>
+            })
+        };
+
+        let config = EncoderConfig {
+            limit: matches.value_of("LIMIT").unwrap().parse().unwrap(),
+            quantizer: matches.value_of("QP").unwrap().parse().unwrap(),
+            speed: matches.value_of("SPEED").unwrap().parse().unwrap(),
+            tune: matches.value_of("TUNE").unwrap().parse().unwrap()
+        };
+
+        // Validate arguments
+        if config.quantizer == 0 {
+            unimplemented!();
+        } else if config.quantizer > 255 || config.speed > 10 {
+            panic!("argument out of range");
+        }
+
+        (io, config)
+    }
+}
 
 /// Encode and write a frame.
 pub fn process_frame(sequence: &mut Sequence, fi: &mut FrameInvariants,
